Events
I setup the frameworks for handling events occurring in the engine.  I
also implement it in a few major places to keep it working.  Potential
improvements can be made to add a "permanent" and "temporary" event
handling system.  Also need to add something to handle events that can
last for more than one frame.