Prioritizing Collision Responses
Made it possible for a dev to create their own collision responses on a
priority-based system.  Will be expanding to make a few more defaults,
but good progress is being made.  This is all managed by the collider so
if you wanted to have different responses for different parts of a mesh
(ie. shot to the head = death) then that's possible now.